extract duplicate formatter methods
