+++ app/models/concerns/componentable.rb
@@ -1,4 +1,6 @@
 module Componentable
+  include DutyExpressionFormattable
+
   SMALL_PRODUCERS_QUOTIENT = 'SPQ'.freeze
   LITERS_OF_PURE_ALCOHOL_QUALIFIER = 'LPA'.freeze
   LITERS_QUALIFIER = 'LTR'.freeze
@@ -106,31 +108,6 @@ module Componentable
 
     def meursing?
       duty_expression_id.in?(DutyExpression::MEURSING_DUTY_EXPRESSION_IDS)
-    end
-
-    def duty_expression_formatter_options
-      {
-        duty_expression_id:,
-        duty_expression_description:,
-        duty_expression_abbreviation:,
-        duty_amount:,
-        monetary_unit: monetary_unit_code,
-        monetary_unit_abbreviation:,
-        measurement_unit:,
-        measurement_unit_qualifier:,
-      }
-    end
-
-    def formatted_duty_expression
-      DutyExpressionFormatter.format(duty_expression_formatter_options.merge(formatted: true))
-    end
-
-    def verbose_duty_expression
-      DutyExpressionFormatter.format(duty_expression_formatter_options.merge(verbose: true))
-    end
-
-    def duty_expression_str
-      DutyExpressionFormatter.format(duty_expression_formatter_options)
     end
   end
 end
